requirements.txt
openai==0.27.8
pyyaml==6.0.1
tqdm==4.65.0
argparse==1.4.0
prettytable==3.8.0
python-dotenv==1.0.0